Remove duplication
Move some more functionality into the `Puzzle` class (like the .NET version) to reduce duplication.
